+++ api/zfs.py
@@ -2,6 +2,7 @@ from flask import request
 from flask_restx import Namespace, Resource, fields
 from flask_jwt_extended import jwt_required
 import subprocess
+from utils.zpool_utils import is_pool_name_exists
 
 zfs_api = Namespace('zfs', description='ZFS 관련 API')
 
@@ -25,23 +26,24 @@ class ZFS_list(Resource):
     @zfs_api.doc(description='zfs 전체 조회')
     @jwt_required()
     def get(self):
-        columns = ['NAME', 'USED', 'AVAIL', 'REFER', 'MOUNTPOINT']
-        result = subprocess.run(['zfs', 'list', '-H', '-o', 'name,used,avail,refer,mountpoint'], capture_output=True, encoding='utf-8')
-        lines = result.stdout.strip().split('\n')
+        try: 
+            columns = ['NAME', 'USED', 'AVAIL', 'REFER', 'MOUNTPOINT']
+            result = subprocess.run(['zfs', 'list', '-H', '-o', 'name,used,avail,refer,mountpoint'], capture_output=True, encoding='utf-8')
+            lines = result.stdout.strip().split('\n')
 
-        zfs_list = []
-        for line in lines:
-            values = line.split('\t')
-            zfs_list.append(dict(zip(columns, values)))
+            zfs_list = []
+            for line in lines:
+                values = line.split('\t')
+                zfs_list.append(dict(zip(columns, values)))
 
-        response = {
-            'zfs': zfs_list,
-            'stderr': result.stderr,
-            'returncode': result.returncode
-        }
+            return {
+                'zfs': zfs_list,
+                'stderr': result.stderr,
+                'returncode': result.returncode
+            }
+        except subprocess.CalledProcessError as e:
+            return {'error': str(e)}, 500    
         
-        return response
-
 # zfs 상세 조회 (속성)
 @zfs_api.route('/properties')
 class ZFS_Status(Resource):
@@ -52,19 +54,28 @@ class ZFS_Status(Resource):
         data = request.get_json()
 
         if not data:
-            return {'error': 'No input data provided'}, 400
+            return {'error': '입력된 데이터가 없습니다.'}, 400
         
         pool_name = data.get('pool_name')
         zfs_name = data.get('zfs_name')
-        
-        key_props = [
-            "type", "creation", "used", "available", "referenced", "mounted", "mountpoint",
-            "compression", "quota", "readonly", "sharenfs", "checksum", "atime", "recordsize", "refreservation"
-        ]
+        full_name = f'{pool_name}/{zfs_name}'
 
-        cmd = ["zfs", "get", ",".join(key_props), f'{pool_name}/{zfs_name}']
-        
         try:
+            # 존재하는 pool인지 확인
+            if not is_pool_name_exists(pool_name):
+                return {'error': f'해당 pool을 찾을 수 없습니다. : {pool_name}'}
+            # 존재하는 zfs인지 확인
+            check = subprocess.run(['zfs', 'list', full_name], capture_output=True, text=True)
+            if check.returncode != 0:
+                return {'error': f'해당 ZFS를 찾을 수 없습니다. : {full_name}', 'stderr':check.stderr}, 400
+            
+            key_props = [
+                "type", "creation", "used", "available", "referenced", "mounted", "mountpoint",
+                "compression", "quota", "readonly", "sharenfs", "checksum", "atime", "recordsize", "refreservation"
+            ]
+
+            cmd = ["zfs", "get", ",".join(key_props), full_name]
+        
             result = subprocess.run(cmd, capture_output=True, text=True)
 
             lines = result.stdout.strip().split("\n")
@@ -84,7 +95,7 @@ class ZFS_Status(Resource):
         
             return response
         except subprocess.CalledProcessError as e:
-            return {'error': str(e)}, 500
+            return {'error': f'속성 조회 중 오류가 발생했습니다: {str(e)}'}, 500
 
 # zfs 생성
 @zfs_api.route('/create')
@@ -94,15 +105,25 @@ class ZFSCreate(Resource):
     @zfs_api.expect(create_zfs_model)
     def post(self):
         data = request.json
-        pool = data.get('pool_name')
-        zfs = data.get('zfs_name')
+        pool_name = data.get('pool_name')
+        zfs_name = data.get('zfs_name')
         full_name = f"{pool}/{zfs}"
 
+        # zfs 이름 규칙
+        # 허용 문자 : 영문자, 숫자, -, _, .
+        # 슬래시(/), 공백, 탭, 특수문자 불가 (pool 이름을 따로 받기 때문에 슬래시 허용 안 하는 걸로)
+        # 대소문자 구분
+        if not bool(re.fullmatch(r'^[a-zA-Z0-9_.-]+$', zfs_name)):
+            return {'error': 'zfs 이름은 영문자, 숫자, "_", "-", "."만 사용할 수 있습니다.'}, 400
+
         try:
+            # 존재하는 pool인지 확인
+            if not is_pool_name_exists(pool_name):
+                return {'error': f'해당 pool을 찾을 수 없습니다. : {pool_name}'}
             # 중복 여부 확인
             check = subprocess.run(['zfs', 'list', full_name], capture_output=True, text=True)
             if check.returncode == 0:
-                return {'error': f'ZFS {full_name} already exists'}, 409
+                return {'error': f'ZFS {full_name}은(는) 이미 존재합니다.'}, 400
             
             # 1. 파일시스템 생성
             subprocess.run(['zfs', 'create', full_name], check=True)
@@ -122,11 +143,11 @@ class ZFSCreate(Resource):
                 subprocess.run(['zfs', 'set', f"mountpoint={data['mountpoint']}", full_name], check=True)
 
             return {
-                'message': f'{full_name} 생성 및 설정 완료'
+                'message': f'{full_name} 생성 및 설정이 완료되었습니다.'
                 }, 201
 
         except subprocess.CalledProcessError as e:
-            return {'error': str(e)}, 500
+            return {'error': f'ZFS 생성 중 오류가 발생했습니다: {str(e)}'}, 500
 
 # zfs 삭제
 @zfs_api.route('/delete/<pool_name>/<zfs_name>')
@@ -136,6 +157,14 @@ class DeleteZFS(Resource):
     def delete(self, pool_name, zfs_name):
         full_name = f'{pool_name}/{zfs_name}'
         try:
+            # 존재하는 pool인지 확인
+            if not is_pool_name_exists(pool_name):
+                return {'error': f'해당 pool을 찾을 수 없습니다. : {pool_name}'}
+            # 존재하는 zfs인지 확인
+            check = subprocess.run(['zfs', 'list', full_name], capture_output=True, text=True)
+            if check.returncode != 0:
+                return {'error': f'해당 ZFS를 찾을 수 없습니다. : {full_name}', 'stderr':check.stderr}, 400
+
             result = subprocess.run(
                 ['zfs', 'destroy', full_name],
                 capture_output=True,
@@ -144,7 +173,7 @@ class DeleteZFS(Resource):
             )
             
             return {
-                'message': f'ZFS: {full_name} deleted successfully',
+                'message': f'ZFS {full_name}가 삭제되었습니다.',
                 'stdout': result.stdout.strip().split('\n'),
                 'stderr': result.stderr,
                 'returncode': result.returncode
@@ -152,7 +181,7 @@ class DeleteZFS(Resource):
             
         except subprocess.CalledProcessError as e:
             return {
-                'error': f'Failed to delete ZFS {full_name}',
+                'error': f'ZFS {full_name} 삭제에 실패하였습니다.',
                 'stdout': e.stdout,
                 'stderr': e.stderr,
                 'returncode': e.returncode
